LeatCode Commit2

+++ src/JavaInterviewProgrammes/FibonacciSeries.java
@@ -6,17 +6,17 @@ public class FibonacciSeries {
     public static void main(String[] args) {
         Scanner sc= new Scanner(System.in);
         System.out.println("Provide the no.of values to be displayed for fibonacci : ");
-        Long n=sc.nextLong();
+        int n=sc.nextInt();
         fibonacciSeriesCreation(n);
         sc.close();
 
     }
-    private static void fibonacciSeriesCreation(Long num){
-        Long first = 0L;
-        Long second= 1L;
-        for(Long i = 1L; i<=num; i++){
+    private static void fibonacciSeriesCreation(int num){
+        int first = 0;
+        int second= 1;
+        for(int i = 1; i<=num; i++){
             System.out.print(first+" ");
-            Long third=first+second;
+            int third=first+second;
             first=second;
             second=third;
         }
